software/requirements.txt
flask>=3.0
pyserial>=3.5
opencv-python-headless>=4.9
